Archivos para el módulo: recuperar contraseña
ACTION_recuperarcon.php:
- Se ha añadido la variable '$cargo_recuperar', la cual obtiene el cargo que el usuario seleccionó para enviar el correo de recuperación de contraseña (line 29).
- Se ha cambiado el condicional para la consulta a la tabla 'tbl_administrador' (line 37).
- Se han realizado dos nuevos condicionales para los cargos de instructor y de personal administrativo (line 64 to 81, and line 91 to 108).
- Se ha cambiado el enlace al formulario para restablecer la contraseña (line 143).

ACTION_restaurarpass.php:
- Se han ingresado 3 nuevos campos input en el formulario (line 93, 129 and 130).
- Se han agregado 3 nuevas consultas para cada cargo del sistema (line 181 to 188, line 191 to 198, and line 201 to 208).

index.php:
- Se ha cambiado el texto de la alerta Bootstrap (line 33).
- Se ha agregado una nueva alerta con código HTML dentro de PHP, la misma muestra al usuario que su contraseña ha sido recuperada con éxito (line 39 to 42).
- Se ha añadido un input de selección al formulario de envío de recuperar contraseña (line 138 to 144).

index.css:
- Se ha añadido el id 'cargo_recuperar', el cual corresponde al input de selección en el formulario de envío de correo para recuperar contraseña (line 71).

// Se han hecho pequeños cambios en los comentarios de los archivos 'system.php', 'helper.php' y 'ACTION_validalogin.php'.

// Se ha eliminado el archivo 'form_cambio.php'.

+++ ACTION_restaurar_pass.php
@@ -1,3 +1,53 @@
+<?php
+
+// Se incluye el archivo 'ACTION_conexionBD.php' para acceder a la conexion de la base de datos 'reda' sin tener que escribir el script de conexión:
+include 'ACTION_conexionBD.php';
+/*
+@var string $key
+@var string $emailone
+@var string $cargo_recuperar
+@var string $curDate
+@var string $query
+@var string $row
+@var string $error
+@var string $expDate
+
+Se verifica si los valores 'token', 'email y 'action' están ya definidos (esto viene desde el enlace que el usuario usa para restaurar su contraseña).
+*/
+// Finalmente, se toma el token, el email del usuario, el cargo del mismo y se verifica si el valor definido 'action' no está vacío:
+if (isset($_GET["key"]) && isset($_GET["email"]) && isset($_GET["cargo"]) && isset($_GET["action"]) && ($_GET["action"] == "reset") && !isset($_POST["action"])) {
+    // Se obtiene el token que corresponde al e-mail del usuario que solicitó el cambio de contraseña en un día y momento determinado:
+    $key = $_GET["key"];
+    // Se obtiene dentro de una variable el e-mail del usuario que solicitó el cambio de contraseña:
+    $emailone = $_GET["email"];
+    // Se obtiene dentro de una variable el cargo que posee el correo electrónico:
+    $cargo_recuperar = $_GET["cargo"];
+    // Se toma la fecha en la que se está abriendo el enlace de recuperación de contraseña:
+    $curDate = date("Y-m-d H:i:s");
+    // Se realiza una consulta a la base de datos 'reda' para verificar si el token abierto existe en la base de datos y si dicho token corresponde a un e-mail existente en la tabla:
+    $query = mysqli_query($con, "SELECT * FROM `password_reset_temp` WHERE `token`='" . $key . "' and `email`='" . $emailone . "';");
+    // Se toman los resultados de la consulta y se guarda en la variable '$row' para realizar un conteo de cuantos resultados existen actualmente en la tabla:
+    $row = mysqli_num_rows($query);
+    // En caso de no existir resultado alguno en la consulta, se ejecuta el siguiente código:
+    if ($row == "") {
+        // Se le avisa al usuario de que el link ha expirado o no es correcto debido a que ya lo ha usado con anterioridad:
+        $error = '<div style="text-align:center; margin-top: 12em; font-size: 1.5em;">
+        <h2>Enlace inválido.</h2>
+        <p>El enlace es invalido o ha expirado. Si no copiaste bien el enlace del correo, o si ya lo has usado (caso en el cual es desactivado),</p>
+        <p><a href="localhost:80/reda/index.php">lo invitamos a dar clic aquí para solicitar un nuevo correo de recuperación para cambiar su contraseña</a></p>
+        </div>';
+        echo $error;
+      // En caso de sí haber un resultado en la base de datos, se ejecuta el siguiente código:  
+    } else {
+        // Se traen los resultados guardados en la variable '$query':
+        $row = mysqli_fetch_assoc($query);
+        // Se trae de la tabla de la base de datos la fila llamada 'expDate':
+        $expDate = $row['expDate'];
+        // Se compara si la fecha de expiración es mayor a la fecha actual:
+        if ($expDate >= $curDate) {
+            // En caso de ser así, el siguiente código HTML con el formulario para recuperar la contraseña será mostrado:
+        ?>
+
 <!DOCTYPE html>
 <!-- Idioma en el que el archivo está definido: -->
 <html lang="es">
@@ -17,53 +67,6 @@
     <!-- Archivo CSS (hoja de estilos) de este archivo en cuestión: -->
     <link rel="stylesheet" type="text/css" href="css/restaurar_pass.css">
 </head>
-
-<?php
-
-// Se incluye el archivo 'ACTION_conexionBD.php' para acceder a la conexion de la base de datos 'reda' sin tener que escribir el script de conexión:
-include ('ACTION_conexionBD.php');
-
-/*
-@var string $key
-@var string $emailone
-@var string $curDate
-@var string $query
-@var string $row
-@var string $error
-@var string $expDate
-
-Se verifica si los valores 'token', 'email y 'action' están ya definidos (esto viene desde el enlace que el usuario usa para restaurar su contraseña).
-*/
-// Finalmente, se toma el token y se verifica si el valor definido 'action' no está vacío:
-if (isset($_GET["token"]) && isset($_GET["email"])
-    && isset($_GET["action"]) && ($_GET["action"] == "reset")
-    && !isset($_POST["action"])) {
-    // Se obtiene el token que corresponde al e-mail del usuario que solicitó el cambio de contraseña en un día y momento determinado:
-    $key = $_GET["token"];
-    // Se obtiene el e-mail del usuario que solicitó el cambio de contraseña:
-    $emailone = $_GET["email"];
-    // Se toma la fecha en la que se está abriendo el enlace de recuperación de contraseña:
-    $curDate = date("Y-m-d H:i:s");
-    // Se realiza una consulta a la base de datos 'reda' para verificar si el token abierto existe en la base de datos y si dicho token corresponde a un e-mail existente en la tabla:
-    $query = mysqli_query($con, "SELECT * FROM `password_reset_temp` WHERE `token`='" . $key . "' and `email`='" . $emailone . "';");
-    // Se toman los resultados de la consulta y se guarda en la variable '$row' para realizar un conteo de cuantos resultados existen actualmente en la tabla:
-    $row = mysqli_num_rows($query);
-    // En caso de no existir resultado alguno en la consulta, se ejecuta el siguiente código:
-    if ($row == "") {
-        // Se le avisa al usuario de que el link ha expirado o no es correcto debido a que ya lo ha usado con anterioridad:
-        $error = '<h2>Link inválido</h2>
-<p>El link es invalido o ha expirado. Si no copiaste bien el link del correo, o ya usaste el link (caso en el cual es desactivado),</p>
-<p><a href="localhost:80/reda-master/index.php">Dé click aquí para solicitar un nuevo correo para cambiar su contraseña</a></p>';
-      // En caso de sí haber un resultado en la base de datos, se ejecuta el siguiente código:  
-    } else {
-        // Se traen los resultados guardados en la variable '$query':
-        $row = mysqli_fetch_assoc($query);
-        // Se trae de la tabla de la base de datos la fila llamada 'expDate':
-        $expDate = $row['expDate'];
-        // Se compara si la fecha de expiración es mayor a la fecha actual:
-        if ($expDate >= $curDate) {
-            // En caso de ser así, el siguiente código HTML con el formulario para recuperar la contraseña será mostrado:
-?>  
 <body>
 <!-- Se colocan los 3 scripts para que la versión 4 de Bootstrap pueda funcionar en el código: -->
 <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
@@ -86,6 +89,8 @@ if (isset($_GET["token"]) && isset($_GET["email"])
                                 <i class='fas fa-sync-alt fa-5x'></i>
                             </div>
                             <p id="font" class="font-weight-bold">Restauración de contraseña</p>
+                            <!-- Se crea un input con un valor 'invisible', lo que significa que el usuario no puede verlo, esto con el objetivo de obtener la acción que el formulario tendrá al momento de ser enviado: -->
+                            <input type="hidden" name="action" value="update">
                             <!-- Se crea el primer campo del formulario (usando la clase 'form-group' de bootstrap para indicar que el campo corresponde a un formulario), el mismo sirve para que el usuario ingrese su contraseña nueva: -->
                             <div class="form-group">
                                     <input type="password" name="new" class="form-control" id="new_pass" placeholder="Digite su nueva contraseña">
@@ -117,13 +122,12 @@ if (isset($_GET["token"]) && isset($_GET["email"])
                                     <!-- Este id también permite la aparición de alertas, en este caso, el id puede indicar al usuario si las contraseñas no coinciden o si la contraseña fue cambiada con éxito: -->
                                     <div id="change2"></div>
                                     <?php
-                                        // Si el valor definido 'success' y el valor obtenido 'success' devuelven true, se muestra otra alerta diferente a la anterior:
-                                        if (isset($_GET["success"]) && $_GET["success"] == 'true') {
-                                        // En este caso, la alerta bootstrap es de tipo 'success' e indica que la contraseña ha sido cambiada con éxito:
-                                        echo "<div class= 'alert alert-success alert-dismissible fade show' role='alert' style='border-radius: 0; font-weight: bold;width: 87%; margin-left: 2em;'><button type='button' class='close' data-dismiss='alert' aria-label='close'><span aria-hidden='true'>&times;</span></button>La contraseña ha sido cambiada con éxito</div>";
-                                        }
+
                                     ?>
                                     <br>
+                                    <!-- Se crean dos inputs invisibles desde los que se guarda el correo electrónico del usuario y el cargo del mismo, esto con el objetivo de identificar, de manera específica, al usuario en cuestión: -->
+                                    <input type="hidden" name="email" value="<?php echo $emailone ?>">
+                                    <input type="hidden" name="cargo" value="<?php echo $cargo_recuperar ?>">
                                     <!-- Se crea un botón con la clase 'btn' de bootstrap. Dicho botón permite guardar los cambios para establecer una nueva contraseña: -->
                                     <button type="submit" name="cambio" class="btn btn-md btn-block text-center" id="enter_button">Guardar cambios
                                     </button>
@@ -135,60 +139,72 @@ if (isset($_GET["token"]) && isset($_GET["email"])
         </div>
 <!-- Aquí finaliza el formulario de restauración de contraseña: -->        
 </form>
-<?php
+<script src="js/cambio_pass.js"></script>
+</body>
+</html>
+
+        <?php
         // Por otro lado, si la fecha actual es mayor que la fecha de expiración, se ejecutará entonces el siguiente código:
-        } else {
+        }else{
             // Se le informará al usuario que el link ha expirado debido a que ha pasado 1 día desde que se solicitó:
-            $error = "<h2>El link ha expirado</h2>
-<p>Estás intentando acceder a un link que fue solicitado hace más de 24 horas. Recuerda que estos códigos son válidos por unicamente un día.<br /><br /></p>";
+            $error = "<div style='text-align:center; margin-top: 12em; font-size: 1.5em;'>
+            <h2>El link ha expirado.</h2>
+            <p>Estás intentando acceder a un enlace que fue solicitado hace más de 24 horas. Recuerda que dichos enlaces son válidos por unicamente un día.<br /><br /></p>
+            </div>";
+            echo $error;
         }
     }
-    // Si el error que debe de mostrarse no se aparece en pantalla, entonces el siguiente código hará que esto pase:
-    if ($error != "") {
-        echo "<div class='error'>" . $error . "</div><br />";
-    }
 }
-
-// Se toma el e-mail del formulario de recuperar contraseña y el valor de 'action' que corresponde a la actualización de la contraseña al darle click al botón de 'Guardar cambios':
-if (isset($_POST["email"]) && isset($_POST["action"]) && ($_POST["action"] == "update")) {
+// Se toma el e-mail del formulario de recuperar , junto al cargo, y el valor de 'action' que corresponde a la actualización de la contraseña al darle click al botón de 'Guardar cambios':
+if (isset($_POST["email"]) &&isset($_POST["cargo"]) && isset($_POST["action"]) && ($_POST["action"] == "update")) {
     // La variable '$error' deja de devolver un mensaje de error:
     $error = "";
     /*
     @var string $pass1
     @var string $pass2
+    @var string $emailone
+    @var string $cargo_recuperar
     @var $con
 
     Se definen 2 variables ('$pass1', '$pass2') las cuales tendrán los datos de los campos de contraseña nueva y de confirmar contraseña, además de permitir con normalidad, usando la función 'mysqli_real_escape_string', que la contraseña nueva sea un string con caracteres especiales:
     */
     $pass1 = mysqli_real_escape_string($con, $_POST["new"]);
     $pass2 = mysqli_real_escape_string($con, $_POST["confirmar_pass"]);
+    // La contraseña digitada en el campo de nueva contraseña será guardada en la base de datos con la encriptación 'hash', usando especificamente la función 'password_hash':
+    $recovery_pass_cifrado = password_hash($pass1, PASSWORD_DEFAULT, array("cost"=>12));
     // Se tiene el e-mail del usuario que solicitó el cambio de contraseña y que anteriormente fue obtenido en este archivo:
     $emailone = $_POST["email"];
-    // Se toma la fecha en la que se está realizando el cambio de contraseña, esto con el objetivo de eliminar el token correcto de la base de datos:
-    $curDate = date("Y-m-d H:i:s");
-    // Si las dos contraseñas ingresadas en los campos no son coincidentes entre sí, se ejecuta entonces el siguiente código:
-    if ($pass1 != $pass2) {
-        // Se le informa al usuario con una alerta de que ambas contraseñas no son coincidentes:
-        echo "<p>Las contraseñas no coinciden.<br /><br /></p>";
+    // Se tiene el cargo del usuario que solicitó el cambio de contraseña y que anteriormente fue obtenido en este archivo:
+    $cargo_recuperar = $_POST["cargo"];
+
+    // Si el cargo del usuario corresponde al de 'administrador', entonces se hará la consulta a la tabla 'tbl_administrador' de la base de datos:
+    if ($cargo_recuperar = "Administrador"){
+        // Se hace una consulta a la base de datos para actualizar la contraseña actual por la nueva ingresada por el usuario ('$recovery_pass_cifrado'). Esta nueva contraseña es establecida al usuario que posee el correo que se ha obtenido anteriormente junto con su cargo específico:
+        mysqli_query($con, "UPDATE `tbl_administrador` SET `contrasena_administrador` = '".$recovery_pass_cifrado."' WHERE `correo_administrador`= '".$emailone."';");
+        // Se hace una consulta a la base de datos para eliminar el token usado por el usuario:
+        mysqli_query($con, "DELETE FROM `password_reset_temp` WHERE `email`= '".$emailone."';");
+        // El usuario es redireccionado al archivo 'index.php' para ser informado, con una alerta tipo 'success' de bootstrap, de que la contraseña ha sido cambiada con éxito:
+        header("location: index.php?password_recovered=true");
     }
-    // Si el error que debe de mostrarse no se aparece en pantalla, entonces el siguiente código hará que esto pase:
-    if ($error != "") {
-        echo "<div class='error'>" . $error . "</div><br />";
-      // De no ser necesario que un error se muestre, el siguiente código es ejecutado:
-    } else {
-        // La contraseña digitada en el campo de nueva contraseña será guardada en la base de datos con la encriptación 'hash', usando especificamente la función 'password_hash':
-        $pass1 = md5($pass1);
-        // Se hace una consulta a la base de datos para actualizar la contraseña actual por la nueva ingresada por el usuario ('$pass1'):
-        mysqli_query($con,
-            "UPDATE `tbl_administrador` SET `contrasena_administrador`='" . $pass1 . "', `trn_date`='" . $curDate . "' WHERE `email`='" . $emailone . "';");
-        // Se elimina automaticamente el token de la base de datos al haber sido usado por el usuario:
-        mysqli_query($con, "DELETE FROM `password_reset_temp` WHERE `email`='" . $emailone . "';");
-        // El usuario es redireccionado al archivo 'ACTION_restaurar_pass.php' para ser informado con una alerta tipo 'success' de bootstrap de que la contraseña ha sido cambiada con éxito:
-        header("location: ACTION_restaurar_pass.php?success=true");
+
+    // Si el cargo del usuario corresponde al de 'instructor', entonces se hará la consulta a la tabla 'tbl_instructor' de la base de datos:
+    if($cargo_recuperar = "Instructor"){
+        // Se hace una consulta a la base de datos para actualizar la contraseña actual por la nueva ingresada por el usuario ('$recovery_pass_cifrado'). Esta nueva contraseña es establecida al usuario que posee el correo que se ha obtenido anteriormente junto con su cargo específico:
+        mysqli_query($con, "UPDATE `tbl_instructor` SET `contrasena_instructor` = '".$recovery_pass_cifrado."' WHERE `correo_instructor`= '".$emailone."';");
+        // Se hace una consulta a la base de datos para eliminar el token usado por el usuario:
+        mysqli_query($con, "DELETE FROM `password_reset_temp` WHERE `email`= '".$emailone."';");
+        // El usuario es redireccionado al archivo 'index.php' para ser informado, con una alerta tipo 'success' de bootstrap, de que la contraseña ha sido cambiada con éxito:
+        header("location: index.php?password_recovered=true");
+    }
+    
+    // Si el cargo del usuario corresponde al de 'personal administrativo', entonces se hará la consulta a la tabla 'tbl_personal_administrativo' de la base de datos:
+    if($cargo_recuperar = "Personal administrativo"){
+        // Se hace una consulta a la base de datos para actualizar la contraseña actual por la nueva ingresada por el usuario ('$recovery_pass_cifrado'). Esta nueva contraseña es establecida al usuario que posee el correo que se ha obtenido anteriormente junto con su cargo específico:
+        mysqli_query($con, "UPDATE `tbl_personal_administrativo` SET `contrasena_administrativo` = '".$recovery_pass_cifrado."' WHERE `correo_administrativo`= '".$emailone."';");
+        // Se hace una consulta a la base de datos para eliminar el token usado por el usuario:
+        mysqli_query($con, "DELETE FROM `password_reset_temp` WHERE `email`= '".$emailone."';");
+        // El usuario es redireccionado al archivo 'index.php' para ser informado, con una alerta tipo 'success' de bootstrap, de que la contraseña ha sido cambiada con éxito:
+        header("location: index.php?password_recovered=true");
     }
 }
 ?>
-<!-- Se añade el uso del archivo 'cambio_pass.js' para permitir la aparición de alertas bootstrap en los casos donde sucedan cambios o anomalías en el formulario: -->
-<script src="js/cambio_pass.js"></script>
-</body>
-</html>
